+++ MoltenVK/MoltenVK/OS/MTLRenderPipelineDescriptor+MoltenVK.m
@@ -20,6 +20,16 @@
 #include "MTLRenderPipelineDescriptor+MoltenVK.h"
 #include "MVKCommonEnvironment.h"
 
+#if MVK_USE_METAL_PRIVATE_API
+// These properties aren't public yet.
+@interface MTLRenderPipelineDescriptor ()
+
+@property(nonatomic, readwrite) NSUInteger sampleMask;
+@property(nonatomic, readwrite) float sampleCoverage;
+
+@end
+#endif
+
 @implementation MTLRenderPipelineDescriptor (MoltenVK)
 
 -(MTLPrimitiveTopologyClass) inputPrimitiveTopologyMVK {
@@ -30,5 +40,16 @@
 -(void) setInputPrimitiveTopologyMVK: (MTLPrimitiveTopologyClass) topology {
 	if ([self respondsToSelector: @selector(setInputPrimitiveTopology:)]) { [self setInputPrimitiveTopology:topology]; }
 }
+
+#if MVK_USE_METAL_PRIVATE_API
+-(NSUInteger) sampleMaskMVK {
+	if ( [self respondsToSelector: @selector(sampleMask)] ) { return self.sampleMask; }
+	return 0xFFFFFFFFFFFFFFFFULL;
+}
+
+-(void) setSampleMaskMVK: (NSUInteger) mask {
+	if ([self respondsToSelector: @selector(setSampleMask:)]) { self.sampleMask = mask; }
+}
+#endif
 
 @end